I created this in visual studio
This may need a few little adjustments I copy and pasted from visual studio without compiling with g++ hopefully it works. I just realized this is averaging the 10^2 10^3 10^4 10^5 instead of 3 separate runs so it needs to be edited i commented where it needs to be change but also the correct files need to be put in the same folder I will still commit this and edit later.